feat(controller): Store artifact repostiory in workflow status. Fixes #6255
