+++ pkg/apis/workflow/v1alpha1/generated.proto
@@ -120,6 +120,27 @@ message ArtifactPaths {
   optional Artifact artifact = 1;
 }
 
+// ArtifactRepository represents an artifact repository in which a controller will store its artifacts
+message ArtifactRepository {
+  // ArchiveLogs enables log archiving
+  optional bool archiveLogs = 1;
+
+  // S3 stores artifact in a S3-compliant object store
+  optional S3ArtifactRepository s3 = 2;
+
+  // Artifactory stores artifacts to JFrog Artifactory
+  optional ArtifactoryArtifactRepository artifactory = 3;
+
+  // HDFS stores artifacts in HDFS
+  optional HDFSArtifactRepository hdfs = 4;
+
+  // OSS stores artifact in a OSS-compliant object store
+  optional OSSArtifactRepository oss = 5;
+
+  // GCS stores artifact in a GCS object store
+  optional GCSArtifactRepository gcs = 6;
+}
+
 // +protobuf.options.(gogoproto.goproto_stringer)=false
 message ArtifactRepositoryRef {
   // The name of the config map. Defaults to "artifact-repositories".
@@ -146,6 +167,14 @@ message ArtifactoryArtifact {
   optional string url = 1;
 
   optional ArtifactoryAuth artifactoryAuth = 2;
+}
+
+// ArtifactoryArtifactRepository defines the controller configuration for an artifactory artifact repository
+message ArtifactoryArtifactRepository {
+  optional ArtifactoryAuth artifactoryAuth = 1;
+
+  // RepoURL is the url for artifactory repo.
+  optional string repoURL = 2;
 }
 
 // ArtifactoryAuth describes the secret selectors required for authenticating to artifactory
@@ -405,6 +434,14 @@ message GCSArtifact {
   optional string key = 2;
 }
 
+// GCSArtifactRepository defines the controller configuration for a GCS artifact repository
+message GCSArtifactRepository {
+  optional GCSBucket gCSBucket = 1;
+
+  // KeyFormat is defines the format of how to store keys. Can reference workflow variables
+  optional string keyFormat = 2;
+}
+
 // GCSBucket contains the access information for interfacring with a GCS bucket
 message GCSBucket {
   // Bucket is the name of the bucket
@@ -460,6 +497,17 @@ message HDFSArtifact {
 
   // Path is a file path in HDFS
   optional string path = 2;
+
+  // Force copies a file forcibly even if it exists (default: false)
+  optional bool force = 3;
+}
+
+// HDFSArtifactRepository defines the controller configuration for an HDFS artifact repository
+message HDFSArtifactRepository {
+  optional HDFSConfig hDFSConfig = 1;
+
+  // PathFormat is defines the format of path to store a file. Can reference workflow variables
+  optional string pathFormat = 2;
 
   // Force copies a file forcibly even if it exists (default: false)
   optional bool force = 3;
@@ -798,6 +846,14 @@ message OSSArtifact {
   optional string key = 2;
 }
 
+// OSSArtifactRepository defines the controller configuration for an OSS artifact repository
+message OSSArtifactRepository {
+  optional OSSBucket oSSBucket = 1;
+
+  // KeyFormat is defines the format of how to store keys. Can reference workflow variables
+  optional string keyFormat = 2;
+}
+
 // OSSBucket contains the access information required for interfacing with an Alibaba Cloud OSS bucket
 message OSSBucket {
   // Endpoint is the hostname of the bucket endpoint
@@ -984,6 +1040,18 @@ message S3Artifact {
 
   // Key is the key in the bucket where the artifact resides
   optional string key = 2;
+}
+
+// S3ArtifactRepository defines the controller configuration for an S3 artifact repository
+message S3ArtifactRepository {
+  optional S3Bucket s3Bucket = 1;
+
+  // KeyFormat is defines the format of how to store keys. Can reference workflow variables
+  optional string keyFormat = 2;
+
+  // KeyPrefix is prefix used as part of the bucket key in which the controller will store artifacts.
+  // DEPRECATED. Use KeyFormat instead
+  optional string keyPrefix = 3;
 }
 
 // S3Bucket contains the access information required for interfacing with an S3 bucket
