+++ workflow/artifactrepositories/mocks/Interface.go
@@ -5,8 +5,6 @@ package mocks
 import (
 	context "context"
 
-	config "github.com/argoproj/argo-workflows/v3/config"
-
 	mock "github.com/stretchr/testify/mock"
 
 	v1alpha1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
@@ -18,15 +16,15 @@ type Interface struct {
 }
 
 // Get provides a mock function with given fields: ctx, ref
-func (_m *Interface) Get(ctx context.Context, ref *v1alpha1.ArtifactRepositoryRefStatus) (*config.ArtifactRepository, error) {
+func (_m *Interface) Get(ctx context.Context, ref *v1alpha1.ArtifactRepositoryRefStatus) (*v1alpha1.ArtifactRepository, error) {
 	ret := _m.Called(ctx, ref)
 
-	var r0 *config.ArtifactRepository
-	if rf, ok := ret.Get(0).(func(context.Context, *v1alpha1.ArtifactRepositoryRefStatus) *config.ArtifactRepository); ok {
+	var r0 *v1alpha1.ArtifactRepository
+	if rf, ok := ret.Get(0).(func(context.Context, *v1alpha1.ArtifactRepositoryRefStatus) *v1alpha1.ArtifactRepository); ok {
 		r0 = rf(ctx, ref)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*config.ArtifactRepository)
+			r0 = ret.Get(0).(*v1alpha1.ArtifactRepository)
 		}
 	}
