+++ workflow/common/common.go
@@ -7,9 +7,6 @@ import (
 )
 
 const (
-	// DefaultArchivePattern is the default pattern when storing artifacts in an archive repository
-	DefaultArchivePattern = "{{workflow.name}}/{{pod.name}}"
-
 	// Container names used in the workflow pod
 	MainContainerName = "main"
 	InitContainerName = "init"
