+++ workflow/controller/workflowpod_test.go
@@ -438,7 +438,7 @@ func TestMetadata(t *testing.T) {
 func TestWorkflowControllerArchiveConfig(t *testing.T) {
 	ctx := context.Background()
 	woc := newWoc()
-	setArtifactRepository(woc.controller, &config.ArtifactRepository{S3: &config.S3ArtifactRepository{
+	setArtifactRepository(woc.controller, &wfv1.ArtifactRepository{S3: &wfv1.S3ArtifactRepository{
 		S3Bucket: wfv1.S3Bucket{
 			Bucket: "foo",
 		},
@@ -450,7 +450,7 @@ func TestWorkflowControllerArchiveConfig(t *testing.T) {
 	assert.Len(t, pods.Items, 1)
 }
 
-func setArtifactRepository(controller *WorkflowController, repo *config.ArtifactRepository) {
+func setArtifactRepository(controller *WorkflowController, repo *wfv1.ArtifactRepository) {
 	controller.artifactRepositories = armocks.DummyArtifactRepositories(repo)
 }
 
@@ -458,7 +458,7 @@ func setArtifactRepository(controller *WorkflowController, repo *config.Artifact
 func TestConditionalNoAddArchiveLocation(t *testing.T) {
 	ctx := context.Background()
 	woc := newWoc()
-	setArtifactRepository(woc.controller, &config.ArtifactRepository{S3: &config.S3ArtifactRepository{
+	setArtifactRepository(woc.controller, &wfv1.ArtifactRepository{S3: &wfv1.S3ArtifactRepository{
 		S3Bucket: wfv1.S3Bucket{
 			Bucket: "foo",
 		},
@@ -478,8 +478,8 @@ func TestConditionalNoAddArchiveLocation(t *testing.T) {
 func TestConditionalAddArchiveLocationArchiveLogs(t *testing.T) {
 	ctx := context.Background()
 	woc := newWoc()
-	setArtifactRepository(woc.controller, &config.ArtifactRepository{
-		S3: &config.S3ArtifactRepository{
+	setArtifactRepository(woc.controller, &wfv1.ArtifactRepository{
+		S3: &wfv1.S3ArtifactRepository{
 			S3Bucket: wfv1.S3Bucket{
 				Bucket: "foo",
 			},
@@ -511,7 +511,7 @@ func TestConditionalArchiveLocation(t *testing.T) {
 		},
 	}
 	woc := newWoc()
-	setArtifactRepository(woc.controller, &config.ArtifactRepository{S3: &config.S3ArtifactRepository{
+	setArtifactRepository(woc.controller, &wfv1.ArtifactRepository{S3: &wfv1.S3ArtifactRepository{
 		S3Bucket: wfv1.S3Bucket{
 			Bucket: "foo",
 		},
